third_party/WebKit/LayoutTests/accessibility/scroll-to-make-visible-nested-expected.txt

Tests that scrolling to make an element visible successfully scrolls multiple nested scrolling views'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

5000-pixel box
5000-pixel box
5000-pixel box
Target
PASS window.pageYOffset is 0
PASS outerContainer.scrollTop is 0
PASS innerContainer.scrollTop is 0
PASS target.getBoundingClientRect().top is >= 15000
PASS window.innerHeight is >= target.getBoundingClientRect().bottom
PASS successfullyParsed is true

TEST COMPLETE
